+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2020-07-03
+
+### Added
+
+- New data set for companies house company data
+
 ## 2020-07-02
 
 ### Added